requirements.txt
langchain
ipykernel
python-dotenv
langchain_community
beautifulsoup4 
requests
pypdf
langchain-openai
chromadb
sentence-transformers
langchain_huggingface
faiss-cpu
streamlit
langchain_groq
langchain_core
fastapi
uvicorn
langserve
faiss-cpu
duckduckgo-search
ddgs